requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2
aiofiles==23.2.0

# Google APIs
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-generativeai==0.3.1

# Logging
structlog==23.2.0
colorlog==6.8.0

# Cache (Optional)
aioredis==2.0.1

# CLI
typer==0.9.0
rich==13.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Security
bandit==1.7.5
safety==2.3.5

# Utilities
python-dotenv==1.0.0
pathlib==1.0.1
typing-extensions==4.8.0
cachetools==5.3.2